*: replace github.com/envoyproxy with github.com/saltosystems
This change replaces the github.com/envoyproxy in favor of
github.com/saltosystems to treat it like our own. The reasoning behind
this is mainly because we wanted to vendor it on our internal building
pipelines. Before this change it wasn't possible as the `go install`
commands considers there is a mismatch between the place from where
we're importing the package and the actual name used to define the
module.