templates/go/message.go: fix paths on the nested messages
When validating a message with nested messages with some paths present,
we are manipulating the paths to send to the nested message validation
the paths without the parent prefix.

This is correct, but the legacy code was not doing it correctly, because:
 1. It was modifying the general path variable
 2. The nested message validation was receiving paths that was not
 belonging to him.

Imagine an InstallationUpdate request, with this update_mask:
 - `digital_key_art.background_image", "last_zone_id"`

On the digital_key_art validation, we would change the paths to:
 - `"background_image", "last_zone_id"`

We had two problems here.
  1. To the digital_key_art validation, we send this two paths to
validate, but the last one (`last_zone_id`) does not make sense.
  2. Next field validations would have the original paths modified.

Even if actually is not causing any problem, it is not logically correct
and we should:
  1. Do not modify the paths variable
  2. Send to the nested validation only the paths that belongs to him.