templates/go/message.go: fix validate whole nested message
When the paths has the whole nested message name it is not performing
any validation. It only works when there is no paths at all or when
specific fields of the nested message are specified.

For instance, an `UpdateInstallation` request to update the
`digital_key_art`'s `background_image`, the validation would be performed
if:
 - there are no update paths
 - the update paths has `digital_key_art.background_image`

But if the update paths has `digital_key_art` the validation would not
be performed.

This PR fixes it forcing to validate all the nested message when it
found its name inside the paths to validate.