templates/go/message.go: fix paths on the nested messages (take2)
Last PR introduces a new bug:
https://github.com/saltosystems/protoc-gen-validate/pull/11

If there are no paths related to the nested message, the validation
is performed with an empty paths, leading to validate all fields.

This PR avoids to perform the validation when there are no paths
related to the nested message.

Example of the generated code:
https://github.com/private-saltoapis/saltoapis-go/pull/502/commits/1f81fcf3b18e926846423a4cb0fe4f589465fdad